refactor(frontend): update core/actions imports to avoid barrel file
