+++ public/app/features/org/state/actions.ts
@@ -1,5 +1,5 @@
 import { getBackendSrv } from '@grafana/runtime';
-import { updateConfigurationSubtitle } from 'app/core/actions';
+import { updateConfigurationSubtitle } from 'app/core/reducers/navModel';
 import { ThunkResult } from 'app/types/store';
 import { UserOrg } from 'app/types/user';
