+++ public/app/features/variables/state/actions.ts
@@ -21,7 +21,7 @@ import {
   VariableWithOptions,
 } from '@grafana/data';
 import { config, locationService, logWarning } from '@grafana/runtime';
-import { notifyApp } from 'app/core/actions';
+import { notifyApp } from 'app/core/reducers/appNotification';
 import { contextSrv } from 'app/core/services/context_srv';
 import { getTimeSrv } from 'app/features/dashboard/services/TimeSrv';
 import { DashboardModel } from 'app/features/dashboard/state/DashboardModel';
